[sse.d.ts] Update `payload` types to support all XMLHttpRequest types
As per the XMLHttpRequest `send()` method's API
https://developer.mozilla.org/en-US/docs/Web/API/XMLHttpRequest/send

The `body` parameter in the `send(body)` method allows for multiple
types, such as (but not limited to) Blobs, ArrayBuffers, and FormData.
The implementation in sse.js only allows for strings to be sent in the
POST request, which is fairly limiting.

This change should allow for all* types supported by the
XMLHttpRequest to be supported via sse.js.

*Omitting TypedArray, as that has 11 different types associated with it,
and I wasn't sure if it was that necessary. This could absolutely be
adjusted to support them.